+++ Backend/data/Shares/Intraday/intradayShare_AAA.csv
@@ -1,0 +1,30 @@
+time,open,high,low,close,volume
+2022-04-07 15:00:00,24.82,24.82,24.8,24.8,2785
+2022-04-07 11:00:00,24.82,24.8202,24.8,24.82,800
+2022-04-06 16:00:00,24.83,24.83,24.83,24.83,105
+2022-04-04 13:00:00,24.85,24.85,24.845,24.85,1207
+2022-04-01 16:00:00,24.78,24.78,24.7799,24.7799,4080
+2022-04-01 15:00:00,24.78,24.78,24.7799,24.7799,4000
+2022-04-01 14:00:00,24.78,24.78,24.78,24.78,100
+2022-04-01 13:00:00,24.77,24.78,24.77,24.78,21723
+2022-03-31 10:00:00,24.8369424954,24.8369424954,24.8369424954,24.8369424954,100
+2022-03-30 10:00:00,24.7969795951,24.7969795951,24.7969795951,24.7969795951,100
+2022-03-28 12:00:00,24.7969795951,24.7969795951,24.7869888701,24.7869888701,493
+2022-03-28 10:00:00,24.7969795951,24.7969795951,24.7969795951,24.7969795951,101
+2022-03-25 10:00:00,24.776998145,24.776998145,24.776998145,24.776998145,325
+2022-03-23 14:00:00,24.7520213324,24.7520213324,24.7520213324,24.7520213324,200
+2022-03-23 10:00:00,24.8069703202,24.8069703202,24.8069703202,24.8069703202,425
+2022-03-21 16:00:00,24.8310479676,24.8310479676,24.8310479676,24.8310479676,120
+2022-03-18 16:00:00,24.8569239455,24.8569239455,24.8569239455,24.8569239455,300
+2022-03-18 14:00:00,24.8023745867,24.8023745867,24.8023745867,24.8023745867,700
+2022-03-18 13:00:00,24.8569239455,24.8569239455,24.8568240382,24.8569239455,900
+2022-03-18 11:00:00,24.851928583,24.851928583,24.851928583,24.851928583,500
+2022-03-17 13:00:00,24.8669146705,24.8669146705,24.8669146705,24.8669146705,400
+2022-03-17 10:00:00,24.8369424954,24.8369424954,24.8369424954,24.8369424954,130
+2022-03-16 16:00:00,24.8769053956,24.8769053956,24.8769053956,24.8769053956,1100
+2022-03-16 13:00:00,24.8669146705,24.8669146705,24.8669146705,24.8669146705,500
+2022-03-15 12:00:00,24.8569239455,24.8569239455,24.8569239455,24.8569239455,1045
+2022-03-14 13:00:00,24.8769053956,24.9118729333,24.8769053956,24.9118729333,640
+2022-03-10 11:00:00,24.8769053956,24.8769053956,24.8769053956,24.8769053956,120
+2022-03-09 12:00:00,24.8769053956,24.8769053956,24.8769053956,24.8769053956,200
+2022-03-09 11:00:00,24.8918914832,24.8918914832,24.8779044681,24.8877952859,401
