+++ Backend/data/Shares/Intraday/intradayShare_AAA.json
@@ -1,0 +1,234 @@
+[
+ {
+  "time": "2022-04-07 15:00:00",
+  "open": "24.82",
+  "high": "24.82",
+  "low": "24.8",
+  "close": "24.8",
+  "volume": "2785"
+ },
+ {
+  "time": "2022-04-07 11:00:00",
+  "open": "24.82",
+  "high": "24.8202",
+  "low": "24.8",
+  "close": "24.82",
+  "volume": "800"
+ },
+ {
+  "time": "2022-04-06 16:00:00",
+  "open": "24.83",
+  "high": "24.83",
+  "low": "24.83",
+  "close": "24.83",
+  "volume": "105"
+ },
+ {
+  "time": "2022-04-04 13:00:00",
+  "open": "24.85",
+  "high": "24.85",
+  "low": "24.845",
+  "close": "24.85",
+  "volume": "1207"
+ },
+ {
+  "time": "2022-04-01 16:00:00",
+  "open": "24.78",
+  "high": "24.78",
+  "low": "24.7799",
+  "close": "24.7799",
+  "volume": "4080"
+ },
+ {
+  "time": "2022-04-01 15:00:00",
+  "open": "24.78",
+  "high": "24.78",
+  "low": "24.7799",
+  "close": "24.7799",
+  "volume": "4000"
+ },
+ {
+  "time": "2022-04-01 14:00:00",
+  "open": "24.78",
+  "high": "24.78",
+  "low": "24.78",
+  "close": "24.78",
+  "volume": "100"
+ },
+ {
+  "time": "2022-04-01 13:00:00",
+  "open": "24.77",
+  "high": "24.78",
+  "low": "24.77",
+  "close": "24.78",
+  "volume": "21723"
+ },
+ {
+  "time": "2022-03-31 10:00:00",
+  "open": "24.8369424954",
+  "high": "24.8369424954",
+  "low": "24.8369424954",
+  "close": "24.8369424954",
+  "volume": "100"
+ },
+ {
+  "time": "2022-03-30 10:00:00",
+  "open": "24.7969795951",
+  "high": "24.7969795951",
+  "low": "24.7969795951",
+  "close": "24.7969795951",
+  "volume": "100"
+ },
+ {
+  "time": "2022-03-28 12:00:00",
+  "open": "24.7969795951",
+  "high": "24.7969795951",
+  "low": "24.7869888701",
+  "close": "24.7869888701",
+  "volume": "493"
+ },
+ {
+  "time": "2022-03-28 10:00:00",
+  "open": "24.7969795951",
+  "high": "24.7969795951",
+  "low": "24.7969795951",
+  "close": "24.7969795951",
+  "volume": "101"
+ },
+ {
+  "time": "2022-03-25 10:00:00",
+  "open": "24.776998145",
+  "high": "24.776998145",
+  "low": "24.776998145",
+  "close": "24.776998145",
+  "volume": "325"
+ },
+ {
+  "time": "2022-03-23 14:00:00",
+  "open": "24.7520213324",
+  "high": "24.7520213324",
+  "low": "24.7520213324",
+  "close": "24.7520213324",
+  "volume": "200"
+ },
+ {
+  "time": "2022-03-23 10:00:00",
+  "open": "24.8069703202",
+  "high": "24.8069703202",
+  "low": "24.8069703202",
+  "close": "24.8069703202",
+  "volume": "425"
+ },
+ {
+  "time": "2022-03-21 16:00:00",
+  "open": "24.8310479676",
+  "high": "24.8310479676",
+  "low": "24.8310479676",
+  "close": "24.8310479676",
+  "volume": "120"
+ },
+ {
+  "time": "2022-03-18 16:00:00",
+  "open": "24.8569239455",
+  "high": "24.8569239455",
+  "low": "24.8569239455",
+  "close": "24.8569239455",
+  "volume": "300"
+ },
+ {
+  "time": "2022-03-18 14:00:00",
+  "open": "24.8023745867",
+  "high": "24.8023745867",
+  "low": "24.8023745867",
+  "close": "24.8023745867",
+  "volume": "700"
+ },
+ {
+  "time": "2022-03-18 13:00:00",
+  "open": "24.8569239455",
+  "high": "24.8569239455",
+  "low": "24.8568240382",
+  "close": "24.8569239455",
+  "volume": "900"
+ },
+ {
+  "time": "2022-03-18 11:00:00",
+  "open": "24.851928583",
+  "high": "24.851928583",
+  "low": "24.851928583",
+  "close": "24.851928583",
+  "volume": "500"
+ },
+ {
+  "time": "2022-03-17 13:00:00",
+  "open": "24.8669146705",
+  "high": "24.8669146705",
+  "low": "24.8669146705",
+  "close": "24.8669146705",
+  "volume": "400"
+ },
+ {
+  "time": "2022-03-17 10:00:00",
+  "open": "24.8369424954",
+  "high": "24.8369424954",
+  "low": "24.8369424954",
+  "close": "24.8369424954",
+  "volume": "130"
+ },
+ {
+  "time": "2022-03-16 16:00:00",
+  "open": "24.8769053956",
+  "high": "24.8769053956",
+  "low": "24.8769053956",
+  "close": "24.8769053956",
+  "volume": "1100"
+ },
+ {
+  "time": "2022-03-16 13:00:00",
+  "open": "24.8669146705",
+  "high": "24.8669146705",
+  "low": "24.8669146705",
+  "close": "24.8669146705",
+  "volume": "500"
+ },
+ {
+  "time": "2022-03-15 12:00:00",
+  "open": "24.8569239455",
+  "high": "24.8569239455",
+  "low": "24.8569239455",
+  "close": "24.8569239455",
+  "volume": "1045"
+ },
+ {
+  "time": "2022-03-14 13:00:00",
+  "open": "24.8769053956",
+  "high": "24.9118729333",
+  "low": "24.8769053956",
+  "close": "24.9118729333",
+  "volume": "640"
+ },
+ {
+  "time": "2022-03-10 11:00:00",
+  "open": "24.8769053956",
+  "high": "24.8769053956",
+  "low": "24.8769053956",
+  "close": "24.8769053956",
+  "volume": "120"
+ },
+ {
+  "time": "2022-03-09 12:00:00",
+  "open": "24.8769053956",
+  "high": "24.8769053956",
+  "low": "24.8769053956",
+  "close": "24.8769053956",
+  "volume": "200"
+ },
+ {
+  "time": "2022-03-09 11:00:00",
+  "open": "24.8918914832",
+  "high": "24.8918914832",
+  "low": "24.8779044681",
+  "close": "24.8877952859",
+  "volume": "401"
+ }
+]
