+++ Backend/module/safeNewSymbol.js
@@ -1,7 +1,9 @@
 const fs = require('fs');
 
+// Check if the stock symbol is already in the list
+// -> If it is listed nothing happens
+// -> If not it will be written into the file
 const saveShareSymbol = (symbol) => {
-    console.log("A new Share symbol will be saved");
     const path = 'data/shareSymbols.txt';
     
     const contents = fs.readFileSync(path, 'utf-8').toString().split("\n");
@@ -9,19 +11,19 @@ const saveShareSymbol = (symbol) => {
     let exists = false;
     for (const element of contents) {
         if(element === symbol){
-            console.log("The Share Symbol already exists");
             exists = true;
             break;
         }
     }
     if(!exists){
         fs.writeFile(path, "\n" + symbol, { flag: 'a+' }, err => {})
-        console.log("A new Share symbol is saved");
     }
 }
 
+// Check if the Crypto symbol is already in the list.
+// -> If it is listed nothing happens
+// -> If not it will be written into the file
 const saveCryptoSymbol = (symbol) => {
-    console.log("A new Crypto symbol will be saved");
     const path = 'data/cryptoSymbols.txt';
 
     const contents = fs.readFileSync(path, 'utf-8').toString().split("\n");
@@ -29,14 +31,12 @@ const saveCryptoSymbol = (symbol) => {
     let exists = false;
     for (const element of contents) {
         if(element === symbol){
-            console.log("The Share Symbol already exists");
             exists = true;
             break;
         }
     }
     if(!exists){
         fs.writeFile(path, "\n" + symbol, { flag: 'a+' }, err => {})
-        console.log("A new Crypto symbol is saved"); 
     }
            
 }
