+++ Backend/package.json
@@ -7,7 +7,7 @@
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "node server.js"
   },
-  "author": "",
+  "author": "Steffen Kruschina @fefifef",
   "license": "ISC",
   "dependencies": {
     "axios": "^0.26.1",
