+++ Backend/routes/Cryptos/intradayCrypto.js
@@ -1,18 +1,19 @@
-const updateDataFromAPI = require('../module/updateDataFromAPI');
-const safeNewSymbol = require('../module/safeNewSymbol');
+const updateDataFromAPI = require('../../module/updateCryptoDataFromAPI');
+const safeNewSymbol = require('../../module/safeNewSymbol');
 
 
 const userRoutes = (app, fs) => {
 
-    app.get('/monthlyShare', (req, res, apiKey) => {
+    app.get('/intradayCrypto', (req, res, apiKey) => {
       if(req.query.symbol){
         const symbol = req.query.symbol;
-        const dataPath = './data/Shares/Monthly/monthlyShare_' + symbol + '.json';
+        const dataPath = './data/Crypto/Intraday/intradayCrypto_' + symbol + '.json';
         
+
         fs.access(dataPath, fs.F_OK, (err) => {
           if (err) {
-            updateDataFromAPI.updateMonthlySeriesShare(symbol,apiKey).then(() => {
-              safeNewSymbol.saveShareSymbol(symbol);
+            updateDataFromAPI.updateIntradaySeriesCrypto(symbol,60, apiKey).then(() => {
+              safeNewSymbol.saveCryptoSymbol(symbol);
               fs.readFile(dataPath, 'utf8', (err, data) => {
                 if (err) {
                     throw err;
@@ -23,8 +24,6 @@ const userRoutes = (app, fs) => {
             });
             
             return;
-          }else{
-            console.log("File exists");
           }
           
           fs.readFile(dataPath, 'utf8', (err, data) => {
@@ -40,7 +39,7 @@ const userRoutes = (app, fs) => {
         res.send("NO Symbol");
       }
       
-      });
+    });
   };
-  
+
   module.exports = userRoutes;
