+++ Backend/routes/Cryptos/weeklyCrypto.js
@@ -1,18 +1,18 @@
-const updateDataFromAPI = require('../module/updateDataFromAPI');
-const safeNewSymbol = require('../module/safeNewSymbol');
+const updateDataFromAPI = require('../../module/updateCryptoDataFromAPI');
+const safeNewSymbol = require('../../module/safeNewSymbol');
 
 
 const userRoutes = (app, fs) => {
 
-    app.get('/dailyCrypto', (req, res, apiKey) => {
+    app.get('/weeklyCrypto', (req, res, apiKey) => {
       if(req.query.symbol){
         const symbol = req.query.symbol;
-        const dataPath = './data/Crypto/Daily/dailyCrypto_' + symbol + '.json';
+        const dataPath = './data/Crypto/Weekly/weeklyCrypto_' + symbol + '.json';
         
         
         fs.access(dataPath, fs.F_OK, (err) => {
           if (err) {
-            updateDataFromAPI.updateDailySeriesCrypto(symbol, apiKey).then(() => {
+            updateDataFromAPI.updateWeeklySeriesCrypto(symbol, apiKey).then(() => {
               safeNewSymbol.saveCryptoSymbol(symbol);
               fs.readFile(dataPath, 'utf8', (err, data) => {
                 if (err) {
@@ -24,8 +24,6 @@ const userRoutes = (app, fs) => {
             });
             
             return;
-          }else{
-            console.log("File exists");
           }
           
           fs.readFile(dataPath, 'utf8', (err, data) => {
