+++ Backend/routes/Shares/monthlyShare.js
@@ -1,18 +1,17 @@
-const updateDataFromAPI = require('../module/updateDataFromAPI');
-const safeNewSymbol = require('../module/safeNewSymbol');
+const updateDataFromAPI = require('../../module/updateShareDataFromAPI');
+const safeNewSymbol = require('../../module/safeNewSymbol');
 
 
 const userRoutes = (app, fs) => {
 
-    app.get('/dailyShare', (req, res, apiKey) => {
+    app.get('/monthlyShare', (req, res, apiKey) => {
       if(req.query.symbol){
         const symbol = req.query.symbol;
-        const dataPath = './data/Shares/Daily/dailyShare_' + symbol + '.json';
+        const dataPath = './data/Shares/Monthly/monthlyShare_' + symbol + '.json';
         
-
         fs.access(dataPath, fs.F_OK, (err) => {
           if (err) {
-            updateDataFromAPI.updateDailySeriesShare(symbol, apiKey).then(() => {
+            updateDataFromAPI.updateMonthlySeriesShare(symbol,apiKey).then(() => {
               safeNewSymbol.saveShareSymbol(symbol);
               fs.readFile(dataPath, 'utf8', (err, data) => {
                 if (err) {
@@ -24,8 +23,6 @@ const userRoutes = (app, fs) => {
             });
             
             return;
-          }else{
-            console.log("File exists");
           }
           
           fs.readFile(dataPath, 'utf8', (err, data) => {
@@ -41,7 +38,7 @@ const userRoutes = (app, fs) => {
         res.send("NO Symbol");
       }
       
-    });
+      });
   };
-
+  
   module.exports = userRoutes;
