+++ Backend/routes/Shares/weeklyShare.js
@@ -1,5 +1,5 @@
-const updateDataFromAPI = require('./../module/updateDataFromAPI');
-const safeNewSymbol = require('./../module/safeNewSymbol');
+const updateDataFromAPI = require('../../module/updateShareDataFromAPI');
+const safeNewSymbol = require('../../module/safeNewSymbol');
 
 
 const userRoutes = (app, fs) => {
@@ -23,8 +23,6 @@ const userRoutes = (app, fs) => {
                 });
                 
                 return;
-              }else{
-                console.log("File exists");
               }
               
               fs.readFile(dataPath, 'utf8', (err, data) => {
