+++ Backend/routes/Watchlist/getCryptoForWatchlist.js
@@ -1,28 +1,27 @@
-const updateDataFromAPI = require('../module/updateDataFromAPI');
-const safeNewSymbol = require('../module/safeNewSymbol');
+const updateDataFromAPI = require('../../module/updateShareDataFromAPI');
+const safeNewSymbol = require('../../module/safeNewSymbol');
 
 
 const userRoutes = (app, fs) => {
 
-    app.get('/getShareForWatchlist', (req, res, apiKey) => {
+    app.get('/getCryptoForWatchlist', (req, res, apiKey) => {
         if(req.query.symbol){
             const symbol = req.query.symbol;
-            const dataPathDailyShare = './data/Shares/Daily/dailyShare_' + symbol + '.json';
-            const dataPathQuotedUSshares = './data/quotedUSshares.json';
+            const dataPathDailyShare = './data/Crypto/Daily/dailyCrypto_' + symbol + '.json';
+            const dataPathQuotedUSshares = './data/quotedCrypto.json';
             const dataPathCurrentCurrency = './data/Currency/currentCurrency.json';
 
 
             fs.access(dataPathDailyShare, fs.F_OK, (err) => {
                 if (err) {
                     updateDataFromAPI.updateDailySeriesShare(symbol, apiKey).then(() => {
-                        safeNewSymbol.saveShareSymbol(symbol);
+                        safeNewSymbol.saveCryptoSymbol(symbol);
                     }).then(() => {
                         setData();
                     });
                     return;
                 }else{
                     setData();
-                    console.log("Daily File exists");
                 }
             });
             const setData = () => {
@@ -49,8 +48,8 @@ const userRoutes = (app, fs) => {
                     dayOneSearchText = dayOneSearchText + dayOne.getDate();
                 }
 
-                let openValue = dailyJson_data['Time Series (Daily)'][dayOneSearchText]['1. open'];
-                let closeValue = dailyJson_data['Time Series (Daily)'][dayOneSearchText]['4. close'];
+                let openValue = dailyJson_data['Time Series (Digital Currency Daily)'][dayOneSearchText]['1a. open (EUR)'];
+                let closeValue = dailyJson_data['Time Series (Digital Currency Daily)'][dayOneSearchText]['4a. close (EUR)'];
 
                 let change = (closeValue - openValue) / closeValue;
                 change = change * 100;
@@ -58,15 +57,12 @@ const userRoutes = (app, fs) => {
 
                 let name;
                 for(let share of quotedUSsharesData){
-                    //  console.log(JSON.stringify(share['symbol']));
                     if(share['symbol'] === symbol){
-                        console.log(share['symbol']);
                         name = share['name'];
                         break;
                     }
                 }
-                let value = closeValue * currency['data']['EUR']['value'];
-                value = value.toFixed(2);
+                let value = closeValue;
                 
                 const back = { "name": name, "symbol": symbol, "value": value, "percentChange": change };
