+++ Backend/routes/routes.js
@@ -1,48 +1,48 @@
 // Shares
-const dailyShareRoute = require('./dailyShare');
-const intradayShareRoute = require('./intradayShare');
-const monthlyShareRoute = require('./monthlyShare');
-const weeklyShareRoute = require('./weeklyShare');
+const dailyShareRoute = require('./Shares/dailyShare');
+const intradayShareRoute = require('./Shares/intradayShare');
+const monthlyShareRoute = require('./Shares/monthlyShare');
+const weeklyShareRoute = require('./Shares/weeklyShare');
 
 //Crypto
-const dailyCryptoRoute = require('./dailyCrypto');
-const intradayCryptoRoute = require('./intradayCrypto');
-const monthlyCryptoRoute = require('./monthlyCrypto');
-const weeklyCryptoRoute = require('./weeklyCrypto');
+const dailyCryptoRoute = require('./Cryptos/dailyCrypto');
+const intradayCryptoRoute = require('./Cryptos/intradayCrypto');
+const monthlyCryptoRoute = require('./Cryptos/monthlyCrypto');
+const weeklyCryptoRoute = require('./Cryptos/weeklyCrypto');
 
 //Allgemein
-const quotedUSshares = require('./quotedUSshares');
-const companyOverviewRoute = require('./companyOverview');
-const search = require('./search');
+const quotedUSshares = require('./Generals/quotedUSshares');
+const companyOverviewRoute = require('./Generals/companyOverview');
+const search = require('./Generals/search');
 
 //Watchlist
-const getShareForWatchlist = require('./getShareForWatchlist');
-const getCryptoForWatchlist = require('./getCryptoForWatchlist');
+const getShareForWatchlist = require('./Watchlist/getShareForWatchlist');
+const getCryptoForWatchlist = require('./Watchlist/getCryptoForWatchlist');
 
-const appRouter = (app, fs, apiKey) => {
+const appRouter = (app, fs, apiKeys) => {
   app.get('/', (req, res) => {
     res.send('welcome to the api-server');
   });
   //Shares
-  dailyShareRoute(app, fs, apiKey);
-  intradayShareRoute(app, fs, apiKey);
-  monthlyShareRoute(app, fs, apiKey);
-  weeklyShareRoute(app, fs, apiKey);
+  dailyShareRoute(app, fs, apiKeys[0]);
+  intradayShareRoute(app, fs, apiKeys[0]);
+  monthlyShareRoute(app, fs, apiKeys[0]);
+  weeklyShareRoute(app, fs, apiKeys[0]);
 
   //Crypto
-  dailyCryptoRoute(app, fs, apiKey);
-  intradayCryptoRoute(app, fs, apiKey);
-  monthlyCryptoRoute(app, fs, apiKey);
-  weeklyCryptoRoute(app, fs, apiKey);
+  dailyCryptoRoute(app, fs, apiKeys[0]);
+  intradayCryptoRoute(app, fs, apiKeys[0]);
+  monthlyCryptoRoute(app, fs, apiKeys[0]);
+  weeklyCryptoRoute(app, fs, apiKeys[0]);
   
   //Allgemein
-  quotedUSshares(app, fs, apiKey);
-  companyOverviewRoute(app, fs, apiKey);
+  quotedUSshares(app, fs, apiKeys[0]);
+  companyOverviewRoute(app, fs, apiKeys[1]);
   search(app, fs);
 
   //Watchlist
-  getShareForWatchlist(app,fs,apiKey);
-  getCryptoForWatchlist(app,fs,apiKey);
+  getShareForWatchlist(app,fs,apiKeys[0]);
+  getCryptoForWatchlist(app,fs,apiKeys[0]);
 };
 
 module.exports = appRouter;
